Attempts to deal with timeouts.
libircclient does not seem to detect any issue when the connection has
been dropped due to external causes. I added a test (getting the current
channel topic)that should trigger if 5min go by without anything happening,
hopefully this will cause libircclient to detect the connection problem, if
not, implement a timeout from the lack of replies concerning the topic
should not be too hard.